feat: follow the same JSON:API structure for errors as the JS template
In the [JS template](https://github.com/mu-semtech/mu-javascript-template/blob/master/helpers/mu/server.js#L30-L35)
errors get a message passed in which is returned as the title member.

Note that the JS template should also pass status as a member of the
error obejct, as happens here, but it currently doesn't.