Allow caching of pip install
This conditionally copies over the requirements.txt.  We know a
Dockerfile will be available during the ONBUILD because otherwise this
image could not be the source image.  The copy is a hack that allows us
to copy over the requirements.txt without crashing if it does not exist.

Installation of the dependencies proceeds before copying over everything
as done beforehand.