feat: don't wrap service code import in a try-except
Using gunicorn's reload mechanism, importing the code in a try-except
meant that the reloader would not pick up changes if the initial code
import threw an error. The reloader checks which files it needs to
listen to notifications for based on its imports. It seems that the
combination of a throwing import inside a try block makes it so the
reloader doesn't register the service entrypoint as a file it needs to
listen to.