chore: set pipeline event conditions at root level
By moving the conditions to the root level, we ensure that the pipelines
are only ran when the applicable conditions are met. The current setup
will run three pipelines no matter what (and clone the repo thrice), but
possibly do nothing afterwards. With the new setup we will e.g. not run
the feature-pipeline when pushing a tag.