Startup the flask dev server the correct way
As indicated in the Flask documentation (https://flask.palletsprojects.com/en/2.3.x/server/#in-code) `app.run` is not the best way to run the flask dev server. It is better to run as `flask run`. This does not crash the server on errors, instead it logs and returns the error and stacktrace.